+++ Web/.bowerrc
@@ -1,0 +1,3 @@
+{
+ "directory":"./plugins"
+}
