+++ Web/bower.json
@@ -1,0 +1,21 @@
+{
+  "name": "Web",
+  "homepage": "https://github.com/ycjcl868/PaintAndGuess_Multi-terminal",
+  "authors": [
+    "信鑫-King <45808948@qq.com>"
+  ],
+  "description": "PaintAndGuess",
+  "main": "",
+  "license": "MIT",
+  "ignore": [
+    "**/.*",
+    "node_modules",
+    "bower_components",
+    "test",
+    "tests"
+  ],
+  "dependencies": {
+    "jquery": "^3.1.1",
+    "bootstrap": "^3.3.7"
+  }
+}
