+++ Web/css/style.less
@@ -1,0 +1,1 @@
+@import "common_fun.less";
